requirements.txt
numpy
requests
pyppeteer
Pillow
